wip composite pages

+++ src/main/java/org/dpolivaev/katas/filesystem/domain/File.java
@@ -1,8 +1,10 @@
 package org.dpolivaev.katas.filesystem.domain;
 
 public interface File extends Element {
-    int size();
+    long size();
     void truncate(long newSize);
-    void write(int offset, int length, byte[] source, long sourceOffset);
-    void read(int offset, int length, byte[] destination, long destinationOffset);
+
+    void write(int offset, long length, byte[] source, long sourceOffset);
+
+    void read(int offset, long length, byte[] destination, long destinationOffset);
 }
